pyproject.toml
[project]
name = "assessment-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "scikit-learn>=1.8.0",
]
